// src/api/service/model/linkMarkDto.ts
/**
 * Things Keeper
 * Will save your notes.
 *
 * OpenAPI spec version: 0.2.0
 * Contact: vrahlinskiy@cinimex.ru
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Tag } from './tag';


export interface LinkMarkDto { 
    created?: Date;
    description?: string;
    id?: string;
    link?: string;
    tags?: Array<Tag>;
    title?: string;
}